Fix generate spec to actually test for overwriting
The bashly.yml that the spec uses does not define a "cli-get" command,
so if the generator *were* overwriting this file the spec would not
catch it. Change the spec to use a command name that the example yml
does have, so that it would be overwritten if this happens.