Move bash 3 version check to top of file for early exit
Bash executes the script as it's reading it. When it encounters a function,
it parses the function body before continuing. The bashly-generated script
has *all* of its code in functions which appear before any code is executed
(the initialize/run functions are only called at the bottom of the script).
Normally this is useful, since it ensures that bash has loaded/parsed the
whole script before beginning execution; however, in the case of this
version check it means that if any of the functions contain syntax that
Bash 3 doesn't understand, it will choke on that and crash (with a much
less helpful error) before it has a chance to execute the version check.

This commit moves that check to the very top of the file, where it will be
executed and trigger an early exit before bash tries to parse the rest of
the script, so that it will work correctly regardless of the remainder of
the contents.