Implement Support for Alternate Dependencies
Resolves #354

The change to `config_validator.rb` is my least favourite part of this pass.
However, it seems cleaner to reimplement the nested portions of
`assert_hash` with specific handling. I suspect that some of this could be
simplified, but I also made the simplifying assumption that if you’re
specifying *one* hash value implementation for dependencies, you should
specify them all that way.

The changes to `run.gtx` are required to prevent shellcheck failure since
`$dependencies` is unused otherwise. This could be changed to stay within
`parse_requirements`, but it would require `declare -gA dependencies` above
`local result` with the same shellcheck or some otherwise mostly-harmless
access of the variable. The `-gA` is required so that the variable is
visible outside of `parse_requirements`, since `declare` inside of
a function is the same as `local`.

The otherwise mostly-harmless access could be something like:

```
> if ((${#dependencies[@]} != {{ dependencies.size }})); then
>   printf "Missing dependencies\n" >&2
> fi
```